LayoutTests/platform/mac/accessibility/container-node-delete-causes-crash-expected.txt

CONSOLE MESSAGE: line 18: TypeError: 'undefined' is not a function (evaluating 'accessibilityController.accessibleElementById("foo")')
Checks to make sure a heap-use-after-free crash doesn't occur when a container node with an associated accessibility object is deleted from the tree. The heap-use-after free was occuring when the AccessibilityObject corresponding to the child of the text node walked up its parent chain in AccessibilityObject::supportsARIALiveRegion but its parent was already deleted.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
Text
